# apps/api/src/openapi/openapi-admin.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/collections": {
      "get": {
        "operationId": "AdminCollectionsController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/AdminCollectionsListResponseDto"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Admin Collections"]
      },
      "post": {
        "operationId": "AdminCollectionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCollectionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Admin Collections"]
      }
    },
    "/api/admin/collections/{id}": {
      "get": {
        "operationId": "AdminCollectionsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCollectionDto"
                }
              }
            }
          },
          "404": {
            "description": "collection_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "code", "message", "payload"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "enum": [404],
                      "example": 404
                    },
                    "code": {
                      "type": "string",
                      "enum": ["collection_not_found"],
                      "example": "collection_not_found"
                    },
                    "message": {
                      "type": "string",
                      "example": "Collection not found"
                    },
                    "payload": {
                      "type": "object",
                      "properties": {
                        "collectionId": {
                          "type": "string",
                          "example": "00000000-0000-4000-8000-000000000001"
                        }
                      },
                      "example": {
                        "collectionId": "00000000-0000-4000-8000-000000000001"
                      }
                    }
                  }
                },
                "examples": {
                  "collection_not_found": {
                    "summary": "Collection not found",
                    "value": {
                      "status": 404,
                      "code": "collection_not_found",
                      "message": "Collection not found",
                      "payload": {
                        "collectionId": "00000000-0000-4000-8000-000000000001"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Admin Collections"]
      },
      "patch": {
        "operationId": "AdminCollectionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCollectionDto"
                }
              }
            }
          },
          "404": {
            "description": "collection_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "code", "message", "payload"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "enum": [404],
                      "example": 404
                    },
                    "code": {
                      "type": "string",
                      "enum": ["collection_not_found"],
                      "example": "collection_not_found"
                    },
                    "message": {
                      "type": "string",
                      "example": "Collection not found"
                    },
                    "payload": {
                      "type": "object",
                      "properties": {
                        "collectionId": {
                          "type": "string",
                          "example": "00000000-0000-4000-8000-000000000001"
                        }
                      },
                      "example": {
                        "collectionId": "00000000-0000-4000-8000-000000000001"
                      }
                    }
                  }
                },
                "examples": {
                  "collection_not_found": {
                    "summary": "Collection not found",
                    "value": {
                      "status": 404,
                      "code": "collection_not_found",
                      "message": "Collection not found",
                      "payload": {
                        "collectionId": "00000000-0000-4000-8000-000000000001"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Admin Collections"]
      }
    },
...
  },
  "info": {
    "title": "my-noodles-api — Admin",
    "description": "",
    "version": "dev",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AdminListMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "currentTotal": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": ["total", "currentTotal", "page", "limit"]
      },
...
      "LocalizedStringDto": {
        "type": "object",
        "properties": {
          "uk": {
            "type": "string",
            "minLength": 1
          },
          "en": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": ["uk", "en"]
      },
...
      "AdminCollectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/LocalizedStringDto"
          },
          "description": {
            "$ref": "#/components/schemas/LocalizedStringDto"
          },
          "longDescription": {
            "$ref": "#/components/schemas/LocalizedStringDto"
          },
          "emoji": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "particles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "heroImage": {
            "type": "string",
            "nullable": true
          },
          "themeKey": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "longDescription",
          "emoji",
          "color",
          "particles",
          "sortOrder",
          "isActive"
        ]
      },
      "AdminCollectionsListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminCollectionDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/AdminListMetaDto"
          }
        },
        "required": ["items", "meta"]
      },
      "CreateCollectionDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "$ref": "#/components/schemas/LocalizedStringDto"
          },
          "description": {
            "$ref": "#/components/schemas/LocalizedStringDto"
          },
          "longDescription": {
            "$ref": "#/components/schemas/LocalizedStringDto"
          },
          "emoji": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "particles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "heroImage": {
            "type": "string",
            "nullable": true
          },
          "themeKey": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "minimum": 0
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "slug",
          "name",
          "description",
          "longDescription",
          "emoji",
          "color",
          "particles",
          "sortOrder",
          "isActive"
        ]
      },
      "UpdateCollectionDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "$ref": "#/components/schemas/LocalizedStringDto"
          },
          "description": {
            "$ref": "#/components/schemas/LocalizedStringDto"
          },
          "longDescription": {
            "$ref": "#/components/schemas/LocalizedStringDto"
          },
          "emoji": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "particles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "heroImage": {
            "type": "string",
            "nullable": true
          },
          "themeKey": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "minimum": 0
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}